"ads2_m12_a1_1: 0/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ assignments/Module 12/m12/ADS-2_Exam-2/Solution.java
@@ -13,7 +13,7 @@ public class Solution {
 			String[] tokens = sc.nextLine().split(" ");
 			Edge e = new Edge(Integer.parseInt(tokens[0]),
 			                  Integer.parseInt(tokens[1]),
-			                  Double.parseDouble(tokens[2]));
+			                  Float.parseFloat(tokens[2]));
 			ewg.addEdge(e);
 		}
 		String caseToGo = sc.nextLine();
